Warn! Changes in default namedef options.
If you've used `namedefs.inf` for defining informations to 86Box v3.0.0.3068, you have to **explicitly** define the `config` section.

It was in time currently to fix such inconsistencies, that now the extended sections are disabled by default, and not partially enabled. Probably the namedefs-override feature is not so widespread right now.